fix(examples): force drop shotgun pump when shotgun is dropped
There was an issue where if the shotgun was held and then the pump
was grabbed and then the shotgun was dropped then the pump would
stay attached to the controller and because it was a joint it would
hang in the air. This has been fixed by force dropping the pump
if the shotgun is dropped.