fix(examples): ensure InputAxisCreator prompt settings are per project
The InputAxisCreator prompt would not display in a new project if an
existing VRTK project had been created locally and the `Do not prompt
again` option had been checked. This is because the key was being saved
into the `EditorPrefs` which meant that the Unity editor would persist
that key across any project.

This fix introduces a GUID check for the `Assets/VRTK` directory to
make sure the key is unique per project. This has an issue that if
the `VRTK.meta` file is deleted, then a new GUID will exist and the
prompt will display again, however it's unlikely the `VRTK.meta` file
will be deleted so it's an acceptable side effect.